feat: improve task filter chip design
