+++ lib/widgets/search/task_filter_icon.dart
@@ -1,0 +1,100 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:lotti/blocs/journal/journal_page_cubit.dart';
+import 'package:lotti/themes/theme.dart';
+import 'package:lotti/widgets/app_bar/journal_sliver_appbar.dart';
+import 'package:lotti/widgets/misc/wolt_modal_config.dart';
+import 'package:lotti/widgets/search/task_category_filter.dart';
+import 'package:lotti/widgets/search/task_list_toggle.dart';
+import 'package:lotti/widgets/search/task_status_filter.dart';
+import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
+import 'package:wolt_modal_sheet/wolt_modal_sheet.dart';
+
+class TaskFilterIcon extends StatelessWidget {
+  const TaskFilterIcon({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final pageIndexNotifier = ValueNotifier(0);
+
+    SliverWoltModalSheetPage page1(
+      BuildContext modalSheetContext,
+      TextTheme textTheme,
+    ) {
+      return WoltModalSheetPage(
+        hasSabGradient: false,
+        topBarTitle: Text('Tasks Filter', style: textTheme.titleSmall),
+        isTopBarLayerAlwaysVisible: true,
+        trailingNavBarWidget: IconButton(
+          padding: const EdgeInsets.all(WoltModalConfig.pagePadding),
+          icon: const Icon(Icons.close),
+          onPressed: Navigator.of(modalSheetContext).pop,
+        ),
+        child: const Padding(
+          padding: EdgeInsets.only(
+            bottom: 30,
+            left: 20,
+            top: 10,
+            right: 20,
+          ),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  JournalFilter(),
+                  SizedBox(width: 10),
+                  TaskListToggle(),
+                ],
+              ),
+              SizedBox(height: 10),
+              TaskStatusFilter(),
+              TaskCategoryFilter(),
+            ],
+          ),
+        ),
+      );
+    }
+
+    return Padding(
+      padding: const EdgeInsets.only(right: 30),
+      child: IconButton(
+        onPressed: () {
+          WoltModalSheet.show<void>(
+            pageIndexNotifier: pageIndexNotifier,
+            context: context,
+            pageListBuilder: (modalSheetContext) {
+              final textTheme = context.textTheme;
+              return [
+                page1(modalSheetContext, textTheme),
+              ];
+            },
+            modalDecorator: (child) {
+              return MultiBlocProvider(
+                providers: [
+                  BlocProvider.value(
+                    value: context.read<JournalPageCubit>(),
+                  ),
+                ],
+                child: child,
+              );
+            },
+            modalTypeBuilder: (context) {
+              final size = MediaQuery.of(context).size.width;
+              if (size < WoltModalConfig.pageBreakpoint) {
+                return WoltModalType.bottomSheet();
+              } else {
+                return WoltModalType.dialog();
+              }
+            },
+            onModalDismissedWithBarrierTap: () {
+              Navigator.of(context).pop();
+            },
+          );
+        },
+        icon: Icon(MdiIcons.filterVariant),
+      ),
+    );
+  }
+}
